@@ -1,3 +1,8 @@
 module.exports = {
-  presets: ["@vue/cli-plugin-babel/preset"],
-};
+	presets: ['@vue/cli-plugin-babel/preset'],
+	plugins: [
+		'transform-class-properties',
+		'@babel/plugin-proposal-nullish-coalescing-operator',
+		'@babel/plugin-proposal-optional-chaining',
+	],
+}
